factor out Game structure to separate file
